Proyecto limpio

+++ src/main/java/pe/edu/upc/code_civitrack/servicesImplement/SolicitudServiceImplement.java
@@ -1,12 +1,14 @@
 package pe.edu.upc.code_civitrack.servicesImplement;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import pe.edu.upc.code_civitrack.entities.Solicitud;
 import pe.edu.upc.code_civitrack.repositories.ISolicitudRepository;
 import pe.edu.upc.code_civitrack.servicesInterfaces.ISolicitudService;
 
 import java.util.List;
 
+@Service
 public class SolicitudServiceImplement implements ISolicitudService {
     @Autowired
     private ISolicitudRepository sR;
@@ -18,17 +20,14 @@ public class SolicitudServiceImplement implements ISolicitudService {
 
     @Override
     public void insert(Solicitud solicitud) {
-
-    }
-
-    @Override
-    public void insert(Solicitud solicitud) {
         sR.save(solicitud);
     }
 
     @Override
     public void delete(int id) {
-        sR.deleteById(id);
+        if (sR.existsById(id)) {
+            sR.deleteById(id);
+        }
     }
 
     @Override
